Added a UserMessage entity class and created a startup process in DashboardApplication

+++ src/main/java/uk/me/eastmans/admin/model/PersonRepository.java
@@ -11,7 +11,7 @@ import javax.transaction.Transactional;
  * Created by markeastman on 03/10/2016.
  */
 @ApplicationScoped
-public class PersonRepository extends CRUDRepository<Person,Integer> {
+public class PersonRepository extends CRUDRepository<Person,Long> {
     public PersonRepository()
     {
         super(Person.class); // I cant find a way to do this in the super class
